DetailansichtController gefixed + Thymeleaf-Anzeige erstellt (funktioniert nicht) (nirom100)

+++ Elite-Verleih/src/main/java/de/hhu/abschlussprojektmeangirlzhackstreetboys/EliteVerleih/dataaccess/Modell/Artikel.java
@@ -16,8 +16,7 @@ public class Artikel {
 
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-
-    private int artikelId;
+    private long artikelId;
 
     private String artikelName;
